Use EventEmitter instead of promises for sniffer
This will hopefully give users better control over turning the sniffer on/off and accessing packets without blocking execution of the rest of the program.

Still TBD: All the coroutines on the server side. For sniffing someone else's traffic, ARP spoofing has to go continously, as well as sniffing, as well as a thread that sends the sniffed packets over to the client. 🚮